guardian/requirements.txt
pytz
django==1.10.5
django-material==0.12.5
django-guardian==1.4.6
django-viewflow-pro=0.13

